# next-client/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WealthFlow API",
        "description": "API documentation for WealthFlow financial tracking application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base URL"
        }
    ],
    "paths": {
...
        "/savings": {
            "get": {
                "summary": "Get all savings for a user",
                "description": "Retrieves all savings records for a specific user",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID of the user to fetch savings for"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Saving"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new saving goal",
                "description": "Adds a new saving goal to the database",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewSaving"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Saving goal created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Saving"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid saving data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update an existing saving goal",
                "description": "Updates an existing saving goal in the database",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSaving"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saving goal updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Saving"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid saving data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Saving goal not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Saving": {
                "type": "object",
                "required": [
                    "id",
                    "userid",
                    "name",
                    "date",
                    "amount",
                    "goal",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "userid": {
                        "type": "integer",
                        "format": "int64",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Emergency Fund"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-01T00:00:00Z"
                    },
                    "amount": {
                        "type": "string",
                        "example": "5000.00"
                    },
                    "goal": {
                        "type": "string",
                        "example": "10000.00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "atRisk"
                        ],
                        "example": "active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Building emergency fund for unexpected expenses"
                    }
                }
            },
            "NewSaving": {
                "type": "object",
                "required": [
                    "userId",
                    "name",
                    "date",
                    "amount",
                    "goal",
                    "status"
                ],
                "properties": {
                    "userId": {
                        "type": "integer",
                        "example": 15
                    },
                    "name": {
                        "type": "string",
                        "example": "Emergency Fund"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-01T00:00:00Z"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 5000.00
                    },
                    "goal": {
                        "type": "number",
                        "format": "float",
                        "example": 10000.00
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "atRisk"
                        ],
                        "example": "active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Building emergency fund for unexpected expenses"
                    }
                }
            },
            "UpdateSaving": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Emergency Fund"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-01T00:00:00Z"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 6000.00
                    },
                    "goal": {
                        "type": "number",
                        "format": "float",
                        "example": 10000.00
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "atRisk"
                        ],
                        "example": "active"
                    },
                    "description": {
                        "type": "string",
                        "example": "Building emergency fund for unexpected expenses"
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "error": {
                        "type": "object",
                        "example": {}
                    }
                }
            }
        }
    }
}